% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tf_idf.R
\name{get_tf_idf}
\alias{get_tf_idf}
\title{Function computes tf_idf values for a collection of text features}
\usage{
get_tf_idf(
  x,
  type = "word",
  n_gram = 2L,
  ngram_delimiter = " ",
  feature_id = "document",
  feature_text = "text",
  stopwords = NULL,
  lowercase = TRUE,
  strip_numeric = FALSE,
  strip_punct = FALSE,
  strip_non_alphanum = FALSE
)
}
\arguments{
\item{x}{a data frame with a column of feature text (by book, chapter, etc.) and a
corresponding column that identifies the source of the text (book name, chapter number,
etc). Each row of text should be a character vector of length of 1.}

\item{type}{A string that sets the type of tokenization. Acceptable values
are \dQuote{character}, \dQuote{word}, \dQuote{sentence}, \dQuote{ngram},
\dQuote{line}, \dQuote{paragraph}.}

\item{n_gram}{The number of words in the n-gram if \code{type} is \dQuote{ngram}.
Must be an integer greater than or equal to 1L.}

\item{ngram_delimiter}{The separator between words in a n-gram if \code{type}
is \dQuote{ngram}.}

\item{feature_id}{The column name from \code{x} that identifies each individual
feature.}

\item{feature_text}{The column name from \code{x} that contains the text for a feature.}

\item{stopwords}{A vector of stop words to be used during tokenization.}

\item{lowercase}{A logical which if TRUE then input text will converted to lower case
with the exception of \dQuote{line} and \dQuote{paragraph}.}

\item{strip_numeric}{A logical which if TRUE input text will be stripped of numbers
if \code{type} is \dQuote{word}.}

\item{strip_punct}{A logical which if TRUE input text will be stripped of punctuation
if \code{type} is \dQuote{word} or \dQuote{sentence}}

\item{strip_non_alphanum}{A logical which if TRUE then input text will be stripped of
non-alphanumerics if \code{type} is \dQuote{character}}
}
\value{
A data frame with columns for the feature id, token value, tf, idf,
and tf-idf
}
\description{
The input data frame consist of a column of text broken down as a
collection of books, chapters, sections, paragraphs, etc. (called \code{feature_text})
and a corresponding \code{feature_id} that identifies the origin of the text.
The function tokenizes the column of text (by word, sentence, ngram, etc) and computes
the term frequency(tf), inverse document frequency(idf) and their product(tf-idf).
}
\author{
Rick Dean
}
