% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_words.R
\name{join_words}
\alias{join_words}
\title{Function joins two character vectors of words}
\usage{
join_words(outer_df, inner_df, key_name = "word")
}
\arguments{
\item{outer_df}{The outer data frame of words}

\item{inner_df}{The inner data frame of words}

\item{key_name}{The column name to use as the common key between
\code{outer_v} and \code{inner_v}}
}
\value{
A data.table of the inner joined words between \code{outer_v} and \code{inner_v}
}
\description{
Function joins two character vectors of words
}
\author{
Rick Dean
}
