% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tf_idf.R
\name{plot_tf_idf}
\alias{plot_tf_idf}
\title{Function plots the top \code{N} ordered tf_idf values}
\usage{
plot_tf_idf(
  tf_idf = NULL,
  N = 15,
  feature_id_val = NULL,
  title = NULL,
  subtitle = NULL,
  x_title = "token",
  y_title = "tf_idf",
  center_titles = FALSE,
  rot_y_tic_label = FALSE,
  bar_fill = NA,
  bar_color = "black",
  bar_alpha = 1,
  bar_size = 1,
  bar_width = NULL,
  y_limits = NULL,
  y_major_breaks = waiver(),
  y_minor_breaks = waiver(),
  y_labels = waiver(),
  axis_text_size = 11,
  do_coord_flip = FALSE,
  bar_labels = FALSE,
  bar_label_size = 4,
  bar_label_color = "black",
  show_major_grids = TRUE,
  show_minor_grids = TRUE
)
}
\arguments{
\item{tf_idf}{the data frame output from RtextminerPkg::get_tf_idf().}

\item{N}{the number of top tf_idf values to bar chart}

\item{feature_id_val}{the feature_id character string for which the top tf_idf values are
to be plotted.}

\item{title}{A string that sets the overall title.}

\item{subtitle}{A string that sets the overall subtitle.}

\item{x_title}{A string that sets the x axis title.}

\item{y_title}{A string that sets the y axis title.}

\item{center_titles}{A logical which if \code{TRUE} centers both the \code{title} and \code{subtitle}.}

\item{rot_y_tic_label}{A logical which if TRUE rotates the y tic labels 90 degrees for enhanced readability.}

\item{bar_fill}{A string that sets the fill color attribute for the bars.}

\item{bar_color}{A string that sets the outline color attribute for the bars.}

\item{bar_alpha}{A numeric that sets the alpha component attribute to \code{bar_color}.}

\item{bar_size}{A numeric that sets the outline thickness attribute of the bars.}

\item{bar_width}{A numeric that sets the width attribute of the bars.}

\item{y_limits}{A numeric 2 element vector that sets the minimum and maximum for the y axis.
Use NA to refer to the existing minimum and maximum.}

\item{y_major_breaks}{A numeric vector or function that defines the exact major tic locations along the y axis.}

\item{y_minor_breaks}{A numeric vector or function that defines the exact minor tic locations along the y axis.}

\item{y_labels}{A character vector with the same length as \code{y_major_breaks}, that labels the major tics.}

\item{axis_text_size}{A numeric that sets the font size along the axis'. Default is 11.}

\item{do_coord_flip}{A logical which if \code{TRUE} will flip the x and y axis'.}

\item{bar_labels}{A logical which if \code{TRUE} will label each bar with its value.}

\item{bar_label_size}{A numeric that sets the size of the bar labels}

\item{bar_label_color}{A string that sets the color of the bar labels}

\item{show_major_grids}{A logical that controls the appearance of major grids.}

\item{show_minor_grids}{A logical that controls the appearance of minor grids.}
}
\value{
A plot object.
}
\description{
Function brings together RtextminerPkg::get_tf_idf() and
RplotterPkg::create_bar_plot() to plot the top \code{N} ordered tf_idf
values from a data frame of text.
}
\author{
Rick Dean
}
